Initial pass at Previous history handling
This change concatenates each selection to enable you to 'undo' a
selection and go to your previous selection. It does NOT support going
back & forth through questions you have already answered in a typical
Previous/Next workflow, because that wouldn't make a ton of sense for
this workflow where each decision impacts what the next step becomes.

Still, this commit is...a bit weird. The UX and styles are not my fave.
The Main component also needs extraction at this point. This is mostly a
fun experiment with converting state to history in React.