Refactor prompt header concept to simply another prompt to display
The CDC protocol does not really distinguish between the header and
standard prompt concepts, nor do the styles. This was an artifact of how
one of the protocol documents was styled and previous styling of this
app.

This change simplifies life so that each step has an array of prompts;
most have one, but those with multiple paragraphs (previously 'prompt
headers') now have multiple prompts in the array. The rendering for this
is handled by a Prompts component to further simplify the rendering
logic in Main.