Extract Step component from Main
Previously the Main component was handling too much rendering logic on
its own.

This extracts a Step component, which represents a single step in the
history array. Each step renders the previous selected response as well
as the prompts and button options for the current step.

This also cleans up some miscellaneous CSS mayhem.